requirements.txt
Django==4.0.5
djangorestframework==3.13.1
psycopg2-binary==2.9.5
sentry-sdk==1.10.1
